Work done on Forms 11, 12, 13
Major work left on Form 13. To Do: Form 12 (Add Topics): I need to write the query that fetches the newly generated TopicId so that I can insert all the books. Form 11 (Topic DataBase): Need to Delete the Search Button and its code, need to figure out the join to filter with books (currently only filtering with name and subject)